ptool/requirements.txt

typer[all]==0.9.0
rich==13.7.0
google-generativeai==0.3.2
pyperclip==1.8.2
python-dotenv==1.0.0
pydantic==2.5.3
prompt_toolkit==3.0.43